Check capability before deleting message

+++ index.php
@@ -39,18 +39,20 @@ if (isguestuser()) {
     throw new moodle_exception('noguest');
 }
 
+$allowpost = has_capability('local/greetings:postmessages', $context);
+$deletepost = has_capability('local/greetings:deleteownmessage', $context);
+$deleteanypost = has_capability('local/greetings:deleteanymessage', $context);
+
 $action = optional_param('action', '', PARAM_TEXT);
 
 if ($action == 'del') {
     $id = required_param('id', PARAM_TEXT);
 
-    // TODO: Confirm before deleting.
-    $DB->delete_records('local_greetings_messages', array('id' => $id));
+    if ($deleteanypost) {
+        // TODO: Confirm before deleting.
+        $DB->delete_records('local_greetings_messages', array('id' => $id));
+    }
 }
-
-$allowpost = has_capability('local/greetings:postmessages', $context);
-$deletepost = has_capability('local/greetings:deleteownmessage', $context);
-$deleteanypost = has_capability('local/greetings:deleteanymessage', $context);
 
 $messageform = new local_greetings_message_form();
